Move ticks related functions to minimalin.c
